% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cds.R
\docType{data}
\name{fifa}
\alias{fifa}
\title{Generate a dataframe of categorical data}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 18207 rows and 4 columns.}
\usage{
fifa
}
\arguments{
\item{n}{Integer to specify the number of rows (current maximum is 334000)}

\item{ncat}{Integer to specify the number of columns (current maximum is 23)}

\item{theme}{Optional. Select of the following: "science", "movies", "books", "comics", "sports", "games".}
}
\description{
Generate a dataframe of categorical data
}
\examples{
# Default is a dataframe with 10 rows and 4 columns
mycds <- cds()
mycds <- cds(n = 900, ncat = 2, )
}
\keyword{datasets}
